# .crmeb/documents/接口文档/公共接口/public_swaggerApi.json
{
  "swagger": "2.0",
  "info": {
    "title": "crmeb_java_public",
    "version": "last",
    "description": "crmeb_java_public"
  },
  "basePath": "/",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/public/wechat/gitlab": {
      "post": {
        "tags": [
          "企业微信消息推送"
        ],
        "summary": "gitlab钩子",
        "consumes": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "企业微信群token",
            "type": "string"
          },
          {
            "name": "root",
            "in": "body",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int64"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string"
                }
              },
              "title": "CommonResult«object»",
              "$$ref": "#/definitions/CommonResult«object»"
            }
          }
        }
      }
    },
    "/api/public/wechat/push": {
      "get": {
        "tags": [
          "企业微信消息推送"
        ],
        "summary": "消息推送",
        "consumes": [
          "text/plain"
        ],
        "parameters": [
          {
            "name": "message",
            "in": "query",
            "required": false,
            "description": "推送消息内容",
            "type": "string"
          },
          {
            "name": "token",
            "in": "query",
            "required": false,
            "description": "企业微信群token",
            "type": "string"
          },
          {
            "name": "raw",
            "in": "body",
            "description": "raw paramter",
            "schema": {
              "type": "string",
              "format": "binary"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "integer",
                  "format": "int64"
                },
                "data": {
                  "type": "object"
                },
                "message": {
                  "type": "string"
                }
              },
              "title": "CommonResult«object»",
              "$$ref": "#/definitions/CommonResult«object»"
            }
          }
        }
      }
    },
...
  }
}